Use python threading (not qt) for opencv tracking
This has nothing to do GUI, so using generic python threads for prcessing
